requirements.txt
facenet_pytorch==2.5.3
numpy==1.23.2
opencv_python==4.6.0.66
pandas==1.5.1
piexif==1.1.3
Pillow==9.5.0
requests==2.28.1
torch==2.0.0
torchvision==0.15.1
